generator: Update stable anchor to 6.9
Apply commit d93859a ("generator: yml: Switch s390 builds on mainline
with LLVM 18+ over to LLVM=1") to the stable anchor, as the fixes for
s390 LLVM=1 were merged in 6.9.

Apply 9a01a2c ("Add support for RISC-V LTO on -next") to the stable
anchor, now that support is available there.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>