workflows: Change cron string to take ownership of failure emails
According to GitHub's documentation [1], notification emails for
scheduled workflow runs go to the user who initially created the
workflow or the user who last updated the cron syntax in the workflow
file.

To ensure emails get sent to someone active on the project, modify the
cron syntax string so that I receive those emails.

[1]: https://docs.github.com/en/actions/monitoring-and-troubleshooting-workflows/notifications-for-workflow-runs

Closes: https://github.com/ClangBuiltLinux/continuous-integration2/issues/755
Signed-off-by: Nathan Chancellor <nathan@kernel.org>