generator: yml: Disable CONFIG_DRM_OMAP with LoongArch configurations
A recent change in linux-next allows this driver to be built with
CONFIG_COMPILE_TEST=y, which causes a couple of -Wframe-larger-than
warnings to show up with ARCH=loongarch. This appears to be an issue in
LLVM, for which I filed an issue. As this code won't actually run on
LoongArch machines, just disable it to avoid having to turn off
CONFIG_WERROR altogether with this configuration.

Link: https://github.com/ClangBuiltLinux/linux/issues/2038
Signed-off-by: Nathan Chancellor <nathan@kernel.org>