patches: Add fix for UBSAN error in android-mainline
android-mainline enables UBSAN by default, which causes the signed
integer overflow sanitizer to become enabled on kernels that contain
commit 557f8c582a9b ("ubsan: Reintroduce signed overflow sanitizer").
As a result, an overflow error from a change in recent kernel versions
becomes fatal.

The patch for this issue is now pending in the maintainer's tree. Apply
it in until it is merged into android-mainline.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>