patches: Apply iio bmi323 change to arm64 and arm64-fixes tree
The arm64 development tree is typically based on -rc3 but this change
did not get merged until 6.12-rc4, so it is needed to avoid disabling
CONFIG_WERROR for this tree.

arm64-fixes is currently based on -rc1, so apply this change for the
same reason, but it may be able to be dropped sooner.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>