Add a step to check if patches apply before invoking TuxSuite
There is no point in spinning up TuxSuite for a build if any patches
that we need to apply are going to fail to do so. At the beginning of a
build, check if the patches apply to the current repository and
reference. This is done by downloading a tarball of the source, running
'git init' and committing the files, then finally running
'git quiltimport' to make sure that that patches still apply. It would
be simpler to use 'git clone --depth=1' but cloning can be expensive
on the server side and these jobs may run at the same time.

Closes: https://github.com/ClangBuiltLinux/continuous-integration2/issues/796
Signed-off-by: Nathan Chancellor <nathan@kernel.org>