Drop android1{3,2}-*
These trees still use a version of clang 12 so it is unlikely that they
will see a compiler upgrade at this point because it could affect the
KMI, so there is little point in testing them with new versions of
clang. This reduces the number of builds per week by 368.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>