generator: yml: Initial mainline build frequency reduction
Halve the number of builds done with LLVM tip of tree and the latest
stable LLVM version per week, as it should not result in too much delay
in finding issues and at this point, a good amount of breakage coming
out of mainline directly (as opposed to coming from -next) is not
actually clang related. This saves 790 builds per week.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>