generator: yml: Reduce build frequency on mainline and tip with older LLVM versions
In general, we want to keep older versions of LLVM versions working but
we do not need to run builds against these trees every day, as it is
becoming more and more rare to experience a particular problem in an
older version of LLVM but not a newer one (unless we missed adding a
workaround). -next frequency is kept daily to continue to try and catch
problems. This saves about 790 builds a week.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>