generator: yml: Disable Hexagon allmodconfig on 5.15
A recent backport exposed Hexagon's lack of readsb() in 5.15. This is
resolved in mainline by switching to the generic IO accessors but
backporting this is not the cleanest due to the HAS_IOPORT rework that
happened gradually over time.

As Hexagon is already a relatively low priority target and it is
happening in an older LTS release, just disable the build. If others
notice and truly care, we can explore addressing this properly at the
source level.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>