scripts/check-logs.py: Print error from TuxSuite better
Currently, we assume if we do not get a "complete" status back from the
build, the build is not completed. Recently however, there have been an
increase in internal TuxSuite infrastructure errors, which are not
obvious upon inspection.

If the build status is "error", print out the status message and exit
early, as we know that the build is done since it never started :)

Signed-off-by: Nathan Chancellor <nathan@kernel.org>