github: Fix internal problem matcher for color ASCII escape codes
The internal warnings/errors problem matcher regex still does not
highlight the messages properly, even though I tested the regex.
Unfortunately, the reason is only obvious when looking at the raw GitHub
logs, which shows the raw color ASCII escape codes that print_red() and
print_yellow() use.

Add the escape codes to the regex to fix the matching, which ends up
making the matcher more robust because it will only highlight the color
messages that the script generates.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>