generator: yml: Reset CONFIG_EFI_SBAT_FILE for Fedora configurations
Fedora recently added a value for CONFIG_EFI_SBAT_FILE, which breaks the
build because that file is not a part of the upstream source tree. Reset
this value with =n so that the build continues to work properly. This is
only relevant for aarch64 and x86_64, who are the only consumers of this
right now.

Link: https://src.fedoraproject.org/rpms/kernel/c/5c8ed2f569619f87c8be4f2fc7db7834615ce7f1
Signed-off-by: Nathan Chancellor <nathan@kernel.org>