scripts/check-logs.py: Treat empty string values as 'n' for config check
Commit 3e2f19e3 ("generator: yml: Reset CONFIG_EFI_SBAT_FILE for Fedora
configurations") used '=n' to reset a string value to its default due to
a limitation in TuxSuite / TuxMake. This causes a mismatch between the
value provided and the value in the config, resulting in an error from
the config check:

  ERROR: CONFIG_EFI_SBAT_FILE=n not found in .config!

If the value is empty, treat it as 'n' for the sake of this check.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>