generator: yml: Disable i386 builds with clang < 17 on 6.18 and newer
Commit e2ffa15b9baa ("kbuild: Disable CC_HAS_ASM_GOTO_OUTPUT on clang <
17") broken i386 by exposing a path that was not built previously (see
the linked CBL issue). There does not appear to be an easy fix for this,
so for now, just disable these builds to keep the matrix green and see
if anyone else actually cares about this broken combination. It should
be easy enough for folks to upgrade their compiler if they rely on
ARCH=i386 building correctly.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>